requirements.txt
# Core MCP Framework
mcp>=1.0.0

# Web Framework and APIs
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# AI and ML
ollama>=0.1.7

# RSS and News Processing
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0

# Social Media APIs
tweepy>=4.14.0  # Twitter API
praw>=7.7.0     # Reddit API

# Exchange APIs
python-binance>=1.0.19
ccxt>=4.1.0

# Configuration and Environment
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Caching and Database
redis>=5.0.1
motor>=3.3.0  # Async MongoDB driver
aiosqlite>=0.19.0

# Utilities
asyncio-throttle>=1.0.2
backoff>=2.2.1
tenacity>=8.2.3

# Monitoring and Logging
structlog>=23.2.0
prometheus-client>=0.19.0
sentry-sdk>=1.40.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0  # For testing async HTTP

# Security
cryptography>=41.0.0
passlib>=1.7.4

# Development Tools
black>=23.0.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0